add Flagship.close method
