+++ flagship/src/api/ITrackingManager.ts
@@ -1,0 +1,11 @@
+import { ITrackingManagerCommon } from './ITrackingManagerCommon'
+
+export interface ITrackingManager extends ITrackingManagerCommon {
+
+    startBatchingLoop():void
+
+    stopBatchingLoop():void
+
+    sendBatch():Promise<void>
+
+  }
