+++ flagship/src/api/TrackingManager.ts
@@ -1,3 +1,4 @@
+import { BatchTriggeredBy } from '../enum/BatchTriggeredBy'
 import { Activate } from '../hit/Activate'
 import { HitAbstract } from '../hit/index'
 import { TrackingManagerAbstract } from './TrackingManagerAbstract'
@@ -9,5 +10,9 @@ export class TrackingManager extends TrackingManagerAbstract {
 
   public async addHit (hit: HitAbstract): Promise<void> {
     await this.strategy.addHit(hit)
+  }
+
+  public async sendBatch (): Promise<void> {
+    await this.strategy.sendBatch(BatchTriggeredBy.Flush)
   }
 }
